@@ -6,6 +6,8 @@ import com.google.android.gms.ads.*
 import com.google.android.gms.ads.formats.UnifiedNativeAd
 import anaxxes.com.weatherFlow.WeatherFlow.click
 import anaxxes.com.weatherFlow.utils.PurchaseUtils
+import android.annotation.SuppressLint
+import com.google.android.gms.ads.interstitial.InterstitialAd
 
 class AdmobUtil(private val context: Context) {
     private var mInterstitialAd: InterstitialAd? = null
@@ -13,6 +15,7 @@ class AdmobUtil(private val context: Context) {
     private var isInterstitialAdLoaded = true
 
 
+    @SuppressLint("MissingPermission")
     fun loadBannerAd(
             container: ViewGroup,
             adSize: AdSize?
@@ -26,6 +29,7 @@ class AdmobUtil(private val context: Context) {
         mAdView.loadAd(adRequest)
     }
 
+    @SuppressLint("MissingPermission")
     fun loadNativeAd(listener: OnNativeAdListener) {
         val builder = AdLoader.Builder(context, Config.nativeId)
         val loader =
@@ -42,12 +46,12 @@ class AdmobUtil(private val context: Context) {
         mInterstitialAdListener = listener
 
         if ((click++) % 5 == 0 && !PurchaseUtils.isPurchased(context)) {
-            if (mInterstitialAd!!.isLoaded) {
-                isInterstitialAdLoaded = true
-                mInterstitialAd!!.show()
-            } else {
-                mInterstitialAdListener?.onAdListener(false, Interstitial.ON_AD_CLOSED)
-            }
+//            if (mInterstitialAd!!.isLoaded) {
+//                isInterstitialAdLoaded = true
+//                mInterstitialAd!!.show()
+//            } else {
+//                mInterstitialAdListener?.onAdListener(false, Interstitial.ON_AD_CLOSED)
+//            }
         }else{
             mInterstitialAdListener?.onAdListener(false, Interstitial.ON_AD_CLOSED)
         }
@@ -55,63 +59,63 @@ class AdmobUtil(private val context: Context) {
     }
 
     fun initInterstitialAd() {
-        mInterstitialAd = InterstitialAd(context)
-        mInterstitialAd!!.adUnitId = Config.interstitialId
-        mInterstitialAd!!.loadAd(AdRequest.Builder().build())
-        mInterstitialAd!!.adListener = object : AdListener() {
-            override fun onAdLoaded() {
-                super.onAdLoaded()
-                mInterstitialAdListener?.onAdListener(
-                        isInterstitialAdLoaded,
-                        Interstitial.ON_AD_LOADED
-                )
-            }
-
-            override fun onAdClicked() {
-                super.onAdClicked()
-                mInterstitialAdListener?.onAdListener(
-                        isInterstitialAdLoaded,
-                        Interstitial.ON_AD_CLICKED
-                )
-            }
-
-            override fun onAdFailedToLoad(i: Int) {
-                super.onAdFailedToLoad(i)
-                mInterstitialAdListener?.onAdListener(
-                        isInterstitialAdLoaded,
-                        Interstitial.ON_AD_FAILED_TO_LOAD
-                )
-            }
-
-            override fun onAdLeftApplication() {
-                super.onAdLeftApplication()
-                mInterstitialAdListener?.onAdListener(
-                        isInterstitialAdLoaded,
-                        Interstitial.ON_AD_LEFT_APP
-                )
-            }
-
-            override fun onAdOpened() {
-                super.onAdOpened()
-                mInterstitialAdListener?.onAdListener(
-                        isInterstitialAdLoaded,
-                        Interstitial.ON_AD_OPENED
-                )
-            }
-
-            override fun onAdClosed() {
-                super.onAdClosed()
-                mInterstitialAd!!.loadAd(AdRequest.Builder().build())
-                mInterstitialAdListener?.onAdListener(
-                        isInterstitialAdLoaded,
-                        Interstitial.ON_AD_CLOSED
-                )
-            }
-        }
+//        mInterstitialAd = InterstitialAd(context)
+//        mInterstitialAd!!.adUnitId = Config.interstitialId
+//        mInterstitialAd!!.loadAd(AdRequest.Builder().build())
+//        mInterstitialAd!!.adListener = object : AdListener() {
+//            override fun onAdLoaded() {
+//                super.onAdLoaded()
+//                mInterstitialAdListener?.onAdListener(
+//                        isInterstitialAdLoaded,
+//                        Interstitial.ON_AD_LOADED
+//                )
+//            }
+//
+//            override fun onAdClicked() {
+//                super.onAdClicked()
+//                mInterstitialAdListener?.onAdListener(
+//                        isInterstitialAdLoaded,
+//                        Interstitial.ON_AD_CLICKED
+//                )
+//            }
+//
+//            override fun onAdFailedToLoad(i: Int) {
+//                super.onAdFailedToLoad(i)
+//                mInterstitialAdListener?.onAdListener(
+//                        isInterstitialAdLoaded,
+//                        Interstitial.ON_AD_FAILED_TO_LOAD
+//                )
+//            }
+//
+//            override fun onAdLeftApplication() {
+//                super.onAdLeftApplication()
+//                mInterstitialAdListener?.onAdListener(
+//                        isInterstitialAdLoaded,
+//                        Interstitial.ON_AD_LEFT_APP
+//                )
+//            }
+//
+//            override fun onAdOpened() {
+//                super.onAdOpened()
+//                mInterstitialAdListener?.onAdListener(
+//                        isInterstitialAdLoaded,
+//                        Interstitial.ON_AD_OPENED
+//                )
+//            }
+//
+//            override fun onAdClosed() {
+//                super.onAdClosed()
+//                mInterstitialAd!!.loadAd(AdRequest.Builder().build())
+//                mInterstitialAdListener?.onAdListener(
+//                        isInterstitialAdLoaded,
+//                        Interstitial.ON_AD_CLOSED
+//                )
+//            }
+//        }
     }
 
 
     init {
-        MobileAds.initialize(context, Config.appId)
+//        MobileAds.initialize(context, Config.appId)
     }
 }
