@@ -61,7 +61,7 @@ public class HourlyListActivity extends GeoActivity {
     }
 
     @Override
-    public View getSnackbarContainer() {
+    public View getSnackBarContainer() {
         return null;
     }
 
