@@ -37,7 +37,7 @@ public class AboutActivity extends GeoActivity
     }
 
     @Override
-    public View getSnackbarContainer() {
+    public View getSnackBarContainer() {
         return container;
     }
 
