@@ -30,7 +30,7 @@ public class SelectProviderActivity extends GeoActivity {
     }
 
     @Override
-    public View getSnackbarContainer() {
+    public View getSnackBarContainer() {
         return findViewById(R.id.activity_settings_container);
     }
 
