@@ -44,7 +44,7 @@ public class AlertActivity extends GeoActivity {
     }
 
     @Override
-    public View getSnackbarContainer() {
+    public View getSnackBarContainer() {
         return container;
     }
 
