@@ -1,38 +1,33 @@
 package anaxxes.com.weatherFlow.ui.adapter;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.core.content.ContextCompat;
-import androidx.recyclerview.widget.GridLayoutManager;
-import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 
 import anaxxes.com.weatherFlow.R;
-import anaxxes.com.weatherFlow.basic.model.location.Location;
 import anaxxes.com.weatherFlow.basic.model.option.unit.CloudCoverUnit;
-import anaxxes.com.weatherFlow.basic.model.option.unit.RelativeHumidityUnit;
 import anaxxes.com.weatherFlow.basic.model.option.unit.SpeedUnit;
 import anaxxes.com.weatherFlow.basic.model.weather.Hourly;
-import anaxxes.com.weatherFlow.models.DailyDetailsModel;
 import anaxxes.com.weatherFlow.models.TodayForecastModel;
 import anaxxes.com.weatherFlow.settings.SettingsOptionManager;
-import anaxxes.com.weatherFlow.utils.MyUtils;
 
 public class HourlyForecastAdapter extends RecyclerView.Adapter<HourlyForecastAdapter.HourlyForecastViewHolder> {
 
     private Context context;
     private ArrayList<Hourly> list = new ArrayList<>();
-    private SettingsOptionManager settingsOptionManager = SettingsOptionManager.getInstance(context);
+    private final SettingsOptionManager settingsOptionManager = SettingsOptionManager.getInstance(context);
 
     public HourlyForecastAdapter(Context context) {
         this.context = context;
@@ -42,7 +37,7 @@ public class HourlyForecastAdapter extends RecyclerView.Adapter<HourlyForecastAd
     @Override
     public HourlyForecastAdapter.HourlyForecastViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
         return new HourlyForecastAdapter.HourlyForecastViewHolder(LayoutInflater.from(parent.getContext())
-                .inflate(R.layout.item_hourly, parent, false));
+                .inflate(R.layout.item_infomation, parent, false));
     }
 
     @Override
@@ -57,8 +52,9 @@ public class HourlyForecastAdapter extends RecyclerView.Adapter<HourlyForecastAd
 
     class HourlyForecastViewHolder extends RecyclerView.ViewHolder {
 
-        private ImageView hourylImage;
-        private TextView tvWeatherText, tvHourlyTime, tvTemp;
+        private ImageView hourylImage, iconBack;
+        private TextView tvWeatherText, tvHourlyTime, tvTemp, precipitation, windChill, cloudCover, windSpeed, humidity, dewPoint, ultravioletIndex, windDirection;
+        private LinearLayout moreInfo, expand;
         private RecyclerView hourlyDetailsList;
 
 
@@ -68,50 +64,63 @@ public class HourlyForecastAdapter extends RecyclerView.Adapter<HourlyForecastAd
             hourylImage = itemView.findViewById(R.id.hourlyImage);
             tvWeatherText = itemView.findViewById(R.id.tvHourlyWeatherText);
             tvHourlyTime = itemView.findViewById(R.id.tvHourlyTime);
-            hourlyDetailsList = itemView.findViewById(R.id.hourlyDetailsList);
+//            hourlyDetailsList = itemView.findViewById(R.id.hourlyDetailsList);
             tvTemp = itemView.findViewById(R.id.tvHourlyTemp);
+            precipitation = itemView.findViewById(R.id.precipitation_hours);
+            windChill = itemView.findViewById(R.id.wind_chill_hours);
+            cloudCover = itemView.findViewById(R.id.cloud_cover_hours);
+            windSpeed = itemView.findViewById(R.id.wind_speed_hour);
+            humidity = itemView.findViewById(R.id.humidity_hours);
+            dewPoint = itemView.findViewById(R.id.dew_point_hours);
+            ultravioletIndex = itemView.findViewById(R.id.ultraviolet_index_hours);
+            windDirection = itemView.findViewById(R.id.wind_dir_hours);
+            moreInfo = itemView.findViewById(R.id.more_info_hourly);
+            iconBack = itemView.findViewById(R.id.icon_expand_hourly);
+            expand = itemView.findViewById(R.id.mostly_cloudy);
         }
 
+        @SuppressLint("SetTextI18n")
         public void setData(Hourly model) {
-
+            model.setExpand(false);
+            moreInfo.setVisibility(View.GONE);
             switch (model.getWeatherCode()) {
                 case CLEAR:
-                    if (MyUtils.isNight()) {
-                        hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.ic_night_clear));
+                    if (!model.isDaylight()) {
+                        hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_moon));
                     } else {
-                        hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.weather_clear_day));
+                        hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_clear));
 
                     }
                     break;
                 case PARTLY_CLOUDY:
-                    if (MyUtils.isNight()) {
-                        hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.weather_partly_cloudy_night));
+                    if (!model.isDaylight()) {
+                        hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_cloudy_moon));
                     } else {
-                        hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.weather_partly_cloudy_day));
+                        hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_partly_cloudy));
                     }
                     break;
                 case CLOUDY:
-                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.weather_partly_cloudy_day));
+                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_sun_cloudy));
 
                     break;
                 case RAIN:
-                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.weather_rain));
+                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_rain));
 
                     break;
                 case SNOW:
-                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.weather_snow));
+                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_snow));
 
                     break;
                 case WIND:
-                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.weather_wind));
+                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_weather_wind));
 
                     break;
                 case FOG:
-                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.weather_fog));
+                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_fog));
 
                     break;
                 case HAZE:
-                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.weather_haze));
+                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_fog));
 
                     break;
                 case SLEET:
@@ -124,7 +133,7 @@ public class HourlyForecastAdapter extends RecyclerView.Adapter<HourlyForecastAd
                     break;
                 case THUNDER:
                 case THUNDERSTORM:
-                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.weather_thunder));
+                    hourylImage.setImageDrawable(ContextCompat.getDrawable(context,R.drawable.img_thunder));
 
 
                     break;
@@ -132,20 +141,46 @@ public class HourlyForecastAdapter extends RecyclerView.Adapter<HourlyForecastAd
             }
 
             tvWeatherText.setText(model.getWeatherText());
-            SimpleDateFormat sdf = new SimpleDateFormat("dd/MM/yyyy hh:mm a");
-            tvHourlyTime.setText(sdf.format(new Date(model.getTime())));
+            @SuppressLint("SimpleDateFormat") SimpleDateFormat sdf = new SimpleDateFormat("hh:mm");
+            tvHourlyTime.setText(model.getHour(context));
             tvTemp.setText(model.getTemperature().getTemperature(context, SettingsOptionManager.getInstance(context).getTemperatureUnit()));
+            precipitation.setText("Precipitation: " + settingsOptionManager.getPrecipitationUnit().getPrecipitationText(
+                    context,
+                    model.getPrecipitationProbability().getTotal()));
+            windChill.setText("Wind chill: " + model.getWind().getLevel());
+            cloudCover.setText("Cloud cover: " + CloudCoverUnit.PERCENT.getCloudCoverText(
+                    model.getCloudCover()
+            ));
+            windSpeed.setText("Wind speed: " + settingsOptionManager.getSpeedUnit().getSpeedText(context, model.getWind().getSpeed()));
+            humidity.setText("Humidity:");
+            dewPoint.setText("Dew point: " + settingsOptionManager.getTemperatureUnit().getTemperatureText(context,
+                    model.getDewPoint()));
+            ultravioletIndex.setText("Ultraviolet index: " + model.getUv().getIndex() + " " + model.getUv().getLevel());
+            windDirection.setText("Wind direction: " + model.getWind().getDirection());
+            expand.setOnClickListener(v -> {
+                if (model.isExpand()){
+                    iconBack.setImageResource(R.drawable.ic_back);
+                    moreInfo.setVisibility(View.GONE);
+                    model.setExpand(false);
+                }
+                else {
+                    iconBack.setImageResource(R.drawable.ic_back_up);
+                    moreInfo.setVisibility(View.VISIBLE);
+                    model.setExpand(true);
+                }
+            });
 
 
             TodayForecastAdapter adapter = new TodayForecastAdapter(context);
-            hourlyDetailsList.setAdapter(adapter);
-            hourlyDetailsList.setLayoutManager(new GridLayoutManager(context,3));
+//            hourlyDetailsList.setAdapter(adapter);
+//            hourlyDetailsList.setLayoutManager(new GridLayoutManager(context,3));
 
             adapter.updateData(getHourlyForecastList(model));
 
         }
     }
 
+    @SuppressLint("NotifyDataSetChanged")
     public void updateData(ArrayList<Hourly> newList) {
         list = newList;
         notifyDataSetChanged();
