@@ -40,7 +40,7 @@ public class LiveWallpaperConfigActivity extends GeoActivity {
     }
 
     @Override
-    public View getSnackbarContainer() {
+    public View getSnackBarContainer() {
         return container;
     }
 
